The RadialGauge and LinearGauge are no longer generic. Remove LinearGaugeFor and RadialGaugeFor methods. Closes #289. Closes #290

+++ wrappers/mvc/src/Kendo.Mvc/UI/Gauge/Fluent/GaugeScaleRangesBuilder.cs
@@ -37,7 +37,7 @@ namespace Kendo.Mvc.UI.Fluent
         /// %&gt;
         /// </code>
         /// </example>        
-        public GaugeScaleRangesBuilder From(decimal from)
+        public GaugeScaleRangesBuilder From(double from)
         {
             item.From = from;
             return this;
@@ -60,7 +60,7 @@ namespace Kendo.Mvc.UI.Fluent
         /// %&gt;
         /// </code>
         /// </example>        
-        public GaugeScaleRangesBuilder To(decimal to)
+        public GaugeScaleRangesBuilder To(double to)
         {
             item.To = to;
             return this;
